# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SRTP Backend API
  version: 0.2.0
servers:
  - url: /
paths:
  /auth/wx/login:
    post:
      operationId: loginWithWechat
      summary: Login with wx.login code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WxLoginRequest'
      responses:
        '200':
          description: Login result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users:
    post:
      operationId: createUser
      summary: Create a user record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      operationId: getUserById
      summary: Get a user by id
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me:
    get:
      operationId: getCurrentUser
      summary: Get current user profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me/profile:
    put:
      operationId: updateCurrentUserProfile
      summary: Update current user profile with synchronous blocked-word validation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me/rooms/created:
    get:
      operationId: listMyCreatedRooms
      summary: List rooms created by current user
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Created rooms
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomCardPage'
  /me/rooms/joined:
    get:
      operationId: listMyJoinedRooms
      summary: List rooms joined by current user
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Joined rooms
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomCardPage'
...
  /rooms:
    get:
      operationId: listRooms
      summary: List rooms
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: sport_type
          schema:
            type: string
        - in: query
          name: campus
          schema:
            type: string
        - in: query
          name: date
          schema:
            type: string
            format: date
        - in: query
          name: time_range
          schema:
            type: string
        - in: query
          name: organization
          schema:
            type: string
        - in: query
          name: level
          schema:
            type: string
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
      responses:
        '200':
          description: Room list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomCardPage'
        '400':
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createRoom
      summary: Create a room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDetail'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/{roomId}:
    get:
      operationId: getRoomById
      summary: Get room detail
      parameters:
        - $ref: '#/components/parameters/RoomIdPath'
      responses:
        '200':
          description: Room found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDetail'
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateRoom
      summary: Update room detail
      parameters:
        - $ref: '#/components/parameters/RoomIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '200':
          description: Updated room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDetail'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/{roomId}/close:
    post:
      operationId: closeRoom
      summary: Close a room
      parameters:
        - $ref: '#/components/parameters/RoomIdPath'
      responses:
        '200':
          description: Room closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDetail'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/join-by-code:
    post:
      operationId: joinRoomByCode
      summary: Join a room by invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRoomByCodeRequest'
      responses:
        '200':
          description: Join result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinRoomResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/{roomId}/join:
    post:
      operationId: joinRoomDirectly
      summary: Join a room directly
      parameters:
        - $ref: '#/components/parameters/RoomIdPath'
      responses:
        '200':
          description: Joined successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinRoomResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/{roomId}/apply:
    post:
      operationId: createJoinRequest
      summary: Create a join request
      parameters:
        - $ref: '#/components/parameters/RoomIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJoinRequestRequest'
      responses:
        '201':
          description: Join request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinRequestResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/{roomId}/approve:
    post:
      operationId: approveJoinRequest
      summary: Approve a join request
      parameters:
        - $ref: '#/components/parameters/RoomIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewJoinRequestRequest'
      responses:
        '200':
          description: Join request approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinRequestResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rooms/{roomId}/reject:
    post:
      operationId: rejectJoinRequest
      summary: Reject a join request
      parameters:
        - $ref: '#/components/parameters/RoomIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewJoinRequestRequest'
      responses:
        '200':
          description: Join request rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinRequestResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    UserIdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    RoomIdPath:
      in: path
      name: roomId
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    PageQuery:
      in: query
      name: page
      schema:
        type: integer
        format: int32
        minimum: 1
    PageSizeQuery:
      in: query
      name: page_size
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
  schemas:
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    CreateUserRequest:
      type: object
      required: [auth_uid]
      properties:
        auth_uid:
          type: string
          minLength: 1
          maxLength: 64
    WxLoginRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 128
        auth_uid:
          type: string
          minLength: 1
          maxLength: 64
    UpdateProfileRequest:
      type: object
      description: Update current user profile. Nickname and bio must pass synchronous blocked-word validation before persistence.
      properties:
        nickname:
          type: string
          maxLength: 30
        avatar_url:
          type: string
          maxLength: 255
        gender:
          type: string
          maxLength: 16
        bio:
          type: string
          maxLength: 255
...
    UserResponse:
      $ref: '#/components/schemas/User'
...
    RoomCardPage:
      type: object
      required: [page, page_size, total, items]
      properties:
        page:
          type: integer
          format: int32
        page_size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/RoomCard'
...